refactor(ui): split both pack modals into mirrored sibling hooks
